moved ImageStorageService and related functionality to the correct AL Layer
