+++ SportMap.AL.Abstractions/Services/IImageStorageService.cs
@@ -1,4 +1,4 @@
-namespace SportMap.DAL.Abstractions
+namespace SportMap.AL.Abstractions.Services
 {
     public interface IImageStorageService
     {
