+++ SportMap.AL/Extensions/DIInitializationExtensions.cs
@@ -4,7 +4,6 @@ using SportMap.AL.UseCases.Feeds;
 using SportMap.AL.UseCases.Images;
 using SportMap.AL.UseCases.Users;
 using SportMap.DAL.Cache;
-using SportMap.AL.Abstractions.Services;
 using SportMap.AL.Common;
 
 namespace SportMap.Al.Extensions
